Moved S3 and Vega logic into separate modules
